5/27/20

+++ edit.php
@@ -15,9 +15,9 @@ $connection->setAttribute(PDO::ATTR_ERRMODE, PDO::FETCH_ASSOC);
 }
 catch (PDOException $e){echo "failed to connect to database, " . $e->getMessage();
 }
-$sql = 'SELECT FirstName, MI, LastName, CompanyName, MailingAddress, City, State, Zip, HomePhone, MobilePhone, Email FROM grower WHERE ID = ?';
+$sql = 'SELECT FirstName, MI, LastName, CompanyName, MailingAddress, City, State, Zip, HomePhone, MobilePhone, Email FROM grower WHERE ID = ? AND UserID = ?';
 $stmt = $connection->prepare($sql);
-$stmt->execute([$_COOKIE['PrimeIDGrower']]);
+$stmt->execute([$_COOKIE['PrimeIDGrower'], $_SESSION['ID']]);
 $arr = $stmt->fetch(PDO::FETCH_NUM);
 echo '
   First Name:<input type="text" value="'.$arr[0].'" name="FirstName"></input>
@@ -35,16 +35,16 @@ echo '
 </form><a href="otherfield.php">fields</a>';
   if (isset($_POST['FirstName'])) {
   if ($_POST['delete'] == "on") {
-    $sql = "DELETE FROM grower WHERE ID = ?";
+    $sql = "DELETE FROM grower WHERE ID = ? AND UserID = ?";
     $stmt = $connection->prepare($sql);
-    $stmt->execute([$_COOKIE['PrimeIDGrower']]);
+    $stmt->execute([$_COOKIE['PrimeIDGrower'], $_SESSION['ID']]);
     $_POST['delete'] == "off";
     header("Location: other.php");
   }
   else {
-  $sql = 'UPDATE grower SET FirstName = ?, MI = ?, LastName = ?, CompanyName = ?, MailingAddress = ?, City = ?, State = ?, Zip = ?, HomePhone = ?, MobilePhone = ?, Email = ? WHERE ID = ?';
+  $sql = 'UPDATE grower SET FirstName = ?, MI = ?, LastName = ?, CompanyName = ?, MailingAddress = ?, City = ?, State = ?, Zip = ?, HomePhone = ?, MobilePhone = ?, Email = ? WHERE ID = ? AND UserID = ?';
   $stmt = $connection->prepare($sql);
-  $stmt->execute([$_POST['FirstName'], $_POST['MI'], $_POST['LastName'], $_POST['CompanyName'], $_POST['MailAdd'], $_POST['City'], $_POST['State'], $_POST['ZIP'], $_POST['Home'], $_POST['Mobile'], $_POST['Email'], $_COOKIE['PrimeIDGrower']]);
+  $stmt->execute([$_POST['FirstName'], $_POST['MI'], $_POST['LastName'], $_POST['CompanyName'], $_POST['MailAdd'], $_POST['City'], $_POST['State'], $_POST['ZIP'], $_POST['Home'], $_POST['Mobile'], $_POST['Email'], $_COOKIE['PrimeIDGrower'], $_SESSION['ID']]);
   header("Location: other.php");
 }
 }
