+++ editField.php
@@ -15,9 +15,9 @@ $connection->setAttribute(PDO::ATTR_ERRMODE, PDO::FETCH_ASSOC);
 }
 catch (PDOException $e){echo "failed to connect to database, " . $e->getMessage();
 }
-$sql = 'SELECT * FROM field WHERE ID = ?';
+$sql = 'SELECT * FROM field WHERE ID = ? AND UserID = ?';
 $stmt = $connection->prepare($sql);
-$stmt->execute([$_COOKIE['PrimeIDField']]);
+$stmt->execute([$_COOKIE['PrimeIDField'], $_SESSION['ID']]);
 $arr = $stmt->fetch(PDO::FETCH_NUM);
 echo '
   Field Name:<input type="text" value="'.$arr[2].'" name="FieldName"></input>
@@ -42,20 +42,23 @@ echo '
   How Was It Killed:<input type="radio" name="HowKilled" value="0"'; if ($arr[19] == 0) {echo ' checked';} echo '>Chemical burn down</input><input type="radio" name="HowKilled" value="1"'; if ($arr[19] == 1) {echo ' checked';} echo '>Plowed or Disked under</input>
   <input type="radio" name="HowKilled" value="2"'; if ($arr[19] == 2) {echo ' checked';} echo '>Harvested</input><input type="radio" name="HowKilled" value="3"'; if ($arr[19] == 3) {echo ' checked';} echo '>Other</input>
   Date:<input type="date" value="'.$arr[20].'" name="DateKilled"></input>
+  Number of years in the last 5 manure was applied:<input type="text" value="'.$arr[22].'" name="Last5"></input>
+  Received manure 8 of last 10 years:<input type="radio" name="8of10" value="0"'; if ($arr[23] == 0) {echo ' checked';} echo '>Yes</input><input type="radio" name="8of10" value="1"'; if ($arr[23] == 1) {echo ' checked';} echo '>No</input>
+  <input type="radio" name="8of10" value="2"'; if ($arr[23] == 2) {echo ' checked';} echo '>Don\'t Know</input>
   <input type="submit"></input><input type="checkbox" name="delete">Delete</input>
 </form><a href="manure.php">Manure</a><br><a href="fertapps.php">Fertilizer Applications</a>';
   if (isset($_POST['FieldName'])) {
   if ($_POST['delete'] == "on") {
-    $sql = "DELETE FROM field WHERE ID = ?";
+    $sql = "DELETE FROM field WHERE ID = ? AND UserID = ?";
     $stmt = $connection->prepare($sql);
-    $stmt->execute([$_COOKIE['PrimeIDField']]);
+    $stmt->execute([$_COOKIE['PrimeIDField'], $_SESSION['ID']]);
     $_POST['delete'] == "off";
     header("Location: other.php");
   }
   else {
-  $sql = 'UPDATE field SET FieldName = ?,  Acres = ?,  County = ?,  Township = ?,  Section = ?,  Quarter = ?,  Tillage = ?,  Plantingdate = ?,  LastYearCrop = ?,  YearsCorn = ?,  Irrigated = ?,  Rotational = ?,  CropYear = ?,  CoverCrop = ?,  DateSeeded = ?,  How = ?,  Ncredits = ?,  HowKilled = ?,  DateKilled = ? WHERE ID = ?';
+  $sql = 'UPDATE field SET FieldName = ?,  Acres = ?,  County = ?,  Township = ?,  Section = ?,  Quarter = ?,  Tillage = ?,  Plantingdate = ?,  LastYearCrop = ?,  YearsCorn = ?,  Irrigated = ?,  Rotational = ?,  CropYear = ?,  CoverCrop = ?,  DateSeeded = ?,  How = ?,  Ncredits = ?,  HowKilled = ?,  DateKilled = ? WHERE ID = ? AND UserID = ?';
   $stmt = $connection->prepare($sql);
-  $stmt->execute([$_POST['FieldName'], $_POST['Acres'], $_POST['County'], $_POST['Township'], $_POST['Section'], $_POST['Quarter'], $_POST['Tillage'], $_POST['Plantingdate'], $_POST['LastYearCrop'], $_POST['YearsCorn'], $_POST['Irrigated'], $_POST['Rotational'], $_POST['CropYear'], $_POST['CoverCrop'], $_POST['DateSeeded'], $_POST['HowSeeded'], $_POST['Ncredits'], $_POST['HowKilled'], $_POST['DateKilled'], $_COOKIE['PrimeIDField']]);
+  $stmt->execute([$_POST['FieldName'], $_POST['Acres'], $_POST['County'], $_POST['Township'], $_POST['Section'], $_POST['Quarter'], $_POST['Tillage'], $_POST['Plantingdate'], $_POST['LastYearCrop'], $_POST['YearsCorn'], $_POST['Irrigated'], $_POST['Rotational'], $_POST['CropYear'], $_POST['CoverCrop'], $_POST['DateSeeded'], $_POST['HowSeeded'], $_POST['Ncredits'], $_POST['HowKilled'], $_POST['DateKilled'], $_COOKIE['PrimeIDField'], $_SESSION['ID']]);
   header("Location: otherfield.php");
   }
 }
